Add FastAPI backend for Varian Eclipse ESAPI integration
Fork VoxTell to expose a REST API that allows Varian Eclipse
(via ESAPI C# scripting) to run AI segmentation on 3D CT volumes.

The backend handles DICOM-to-NIfTI conversion, async inference
scheduling, and returns LPS contour points directly consumable
by Eclipse's AddContourOnImagePlane API.

New:
- api/ — FastAPI app with session management, slice-by-slice
  volume upload, background inference jobs, and contour extraction
- pyproject.toml [api] extra for backend dependencies
- .env.example, run.sh, download_model.py for deployment
- Load text encoder in float16 to reduce VRAM usage

Removed:
- uv.lock (unused)

+++ voxtell/inference/predictor.py
@@ -66,7 +66,9 @@ class VoxTellPredictor:
 
         # Embedding model
         self.tokenizer = AutoTokenizer.from_pretrained(text_encoding_model, padding_side='left')
-        self.text_backbone = AutoModel.from_pretrained(text_encoding_model).eval()
+        self.text_backbone = AutoModel.from_pretrained(
+            text_encoding_model, torch_dtype=torch.float16
+        ).eval()
         self.max_text_length = 8192
 
         # Load network settings
